Updates to Antenna code
Added a "top two" grouping
Added detach/attach code around the antenna servos so that we stop
sending pulses to the servo once it has reached the target position.

This will help to reduce the wear on the servos and prolong life.
IT should also mean that we're less likely to damage the servo in a
"crash" as the servos will most likely not be powered.